# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.DocumentListResponse:
    properties:
      documents:
        items:
          $ref: '#/definitions/dto.DocumentUploadResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.DocumentUploadResponse:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  dto.ErrorDetail:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorDetail'
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  dto.ParseDocumentResponse:
    properties:
      id:
        type: string
      parsed_text:
        type: string
      status:
        type: string
      text_preview:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@testgen.com
    name: API Support
  description: API for test generation system with LLM integration and Moodle export
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Test Generation System API
  version: "1.0"
paths:
...
  /documents:
    get:
      description: Get paginated list of documents uploaded by the current user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's documents
      tags:
      - documents
    post:
      consumes:
      - multipart/form-data
      description: Upload a document file (PDF, DOCX, PPTX, TXT) for processing
      parameters:
      - description: Document file
        in: formData
        name: file
        required: true
        type: file
      - description: Document title
        in: formData
        name: title
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DocumentUploadResponse'
        "400":
          description: Invalid input or unsupported file type
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a document
      tags:
      - documents
  /documents/{id}:
    delete:
      description: Delete a document and its associated file
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid document ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a document
      tags:
      - documents
    get:
      description: Get details of a specific document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DocumentUploadResponse'
        "400":
          description: Invalid document ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get document by ID
      tags:
      - documents
  /documents/{id}/parse:
    post:
      description: Extract text content from an uploaded document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ParseDocumentResponse'
        "400":
          description: Invalid document ID or file type
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Document not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Parsing failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Parse a document
      tags:
      - documents
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"